% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_formatting.R
\name{build_prior_qualification_title}
\alias{build_prior_qualification_title}
\title{Build prior qualification title text}
\usage{
build_prior_qualification_title(date, version)
}
\arguments{
\item{date}{Prior qualification date}

\item{version}{Prior qualified version}
}
\value{
Character string for tooltip
}
\description{
Internal function to create prior qualification badge tooltip text.
}
\keyword{internal}
