% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_formatting.R
\name{build_qualification_title}
\alias{build_qualification_title}
\title{Build qualification title text}
\usage{
build_qualification_title(date)
}
\arguments{
\item{date}{Qualification date}
}
\value{
Character string for tooltip
}
\description{
Internal function to create qualification badge tooltip text.
}
\keyword{internal}
