% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_formatting.R
\name{count_releases_last_n_days}
\alias{count_releases_last_n_days}
\title{Count releases from last N days}
\usage{
count_releases_last_n_days(releases, days)
}
\arguments{
\item{releases}{List of release objects from \code{fetch_releases()}}

\item{days}{Number of days to look back from today}
}
\value{
Integer count of qualifying releases.
}
\description{
Internal function to count the number of non-draft, non-prerelease
releases published in the last N days (inclusive).
}
\keyword{internal}
