% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_formatting.R
\name{count_ytd_releases}
\alias{count_ytd_releases}
\title{Count releases from last 365 days}
\usage{
count_ytd_releases(releases)
}
\arguments{
\item{releases}{List of release objects from \code{fetch_releases()}}
}
\value{
Integer count of qualifying releases.
}
\description{
Internal function to count the number of non-draft, non-prerelease
releases published in the last 365 days (inclusive).
}
\keyword{internal}
