% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualification_registry.R
\name{default_qualification_registry_url}
\alias{default_qualification_registry_url}
\title{Get default qualification registry URL}
\usage{
default_qualification_registry_url()
}
\value{
Character string with registry URL
}
\description{
Returns the default URL for the qualification registry, checking options and
environment variables for overrides.
}
\keyword{internal}
