% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_formatting.R
\name{format_milestone_summary}
\alias{format_milestone_summary}
\title{Format milestone summary as HTML}
\usage{
format_milestone_summary(owner, repo, milestones)
}
\arguments{
\item{owner}{Repository owner (GitHub username or organization)}

\item{repo}{Repository name}

\item{milestones}{List of milestone objects from GitHub API}
}
\value{
Character string with formatted HTML
}
\description{
Internal function to format GitHub milestones as styled HTML.
}
\keyword{internal}
