% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_api.R
\name{has_status_code}
\alias{has_status_code}
\title{Check if error has specific HTTP status code}
\usage{
has_status_code(err, code)
}
\arguments{
\item{err}{Error object from GitHub API}

\item{code}{HTTP status code to check for}
}
\value{
Logical indicating if error has the specified status code
}
\description{
Internal function to check if a GitHub API error has a specific status code.
}
\keyword{internal}
