% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_validation.R
\name{or_pipe}
\alias{or_pipe}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{Left-hand side value}

\item{rhs}{Right-hand side value (default if lhs is NULL)}
}
\value{
lhs if not NULL, otherwise rhs
}
\description{
Returns the right-hand side if left-hand side is NULL, otherwise returns
left-hand side.
}
\keyword{internal}
