% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pr_activity_by_user.R
\name{read_pr_review_cache}
\alias{read_pr_review_cache}
\title{Read pull request review cache}
\usage{
read_pr_review_cache(path)
}
\arguments{
\item{path}{Cache file path}
}
\value{
Cached review list or NULL
}
\description{
Read pull request review cache
}
\keyword{internal}
