% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.dash-package.R
\docType{package}
\name{gh.dash}
\alias{gh.dash-package}
\alias{gh.dash}
\title{gh.dash: Dashboards for GitHub Packages}
\description{
Tools for building and maintaining dashboard-focused reports for GitHub repositories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Gilead-BioStats/gh.dash}
  \item Report bugs at \url{https://github.com/Gilead-BioStats/gh.dash/issues}
}

}
\author{
\strong{Maintainer}: Jeremy Wildfire \email{jwildfire@gmail.com}

Authors:
\itemize{
  \item Laura Maxwell \email{laura.maxwell@atorusresearch.com}
}

Other contributors:
\itemize{
  \item Gilead Sciences [copyright holder]
}

}
\keyword{internal}
